Socket.io text adventure idea
Not so much code as it is the beginnings of code.
Any advice will be appreciated.
The player and game are mediated with a server that passes text-based data negotiations.
If that makes any sense. Some portion of the game function must be handled by the server 
This, for example should not work: client:socketC:socketS:telnet because of securtiy. 
Read more in the text file, and yes I am fully clueless when it comes to webservers and sockets.